アップロードモードの追加、各種調整

+++ main.py
@@ -5,6 +5,7 @@ import json
 from multiprocessing import Process
 from tools import Tools
 from thread import Task
+from record import Record
 
 # Reading settings from config file
 config = configparser.ConfigParser()
@@ -21,22 +22,35 @@ task = None
 
 def nextButton(channel):
     if task != None:
+        print("aaaaa")
         task.isContinue = False
 
 while True:
-    req = request.Request(url)
-    with request.urlopen(req) as res:
-        body = json.loads(res.read().decode('utf8'))
-
-    # 認証失敗またはリクエストなし
-    if body == None:
-        continue
-
-    if p != None:
-        p.terminate()
     tools = Tools(nextButton) 
-    task = Task(body, tools)
-    p = Process(target=task.execute())
-    p.start()
-    p.join()
+    print(tools.isButton())
+    if tools.isButton() == 1:
+        tools.TTS("レシピアップロードモードだお")
+        task = Task({}, tools)
+        task.record()
+        print(json)
+        
+    else:
+        tools.TTS("レシピダウンロードモードだお")
+        time.sleep(3)
+        req = request.Request(url)
+        with request.urlopen(req) as res:
+            body = json.loads(res.read().decode('utf8'))
+
+        # 認証失敗またはリクエストなし
+        print(res)
+        if body == None:
+            continue
+
+        if p != None:
+            p.terminate()
+        print(tools.isButton())
+        task = Task(body, tools)
+        p = Process(target=task.execute())
+        p.start()
+        p.join()
